Refactor RGB to HSL conversion to use 16-bit hue and update function signatures for consistency

+++ src/fauxmoESP.h
@@ -150,6 +150,6 @@ class fauxmoESP {
         String _byte2hex(uint8_t zahl);
         String _makeMD5(String text);
         byte* _hs2rgb(uint16_t hue, uint8_t sat);
-        byte* _rgb2hs(byte r, byte g, byte b);
+        uint16_t* _rgb2hs(byte r, byte g, byte b);
         byte* _ct2rgb(uint16_t ct);
 };
